Finish moving Class, Length, Tag and Header to separate files
