+++ src/ber/class.rs
@@ -1,3 +1,6 @@
+use core::convert::TryFrom;
+use core::fmt;
+
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
 pub struct ClassFromIntError(pub(crate) ());
 
@@ -9,4 +12,31 @@ pub enum Class {
     Application = 0b01,
     ContextSpecific = 0b10,
     Private = 0b11,
+}
+
+impl fmt::Display for Class {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let s = match self {
+            Class::Universal => "UNIVERSAL",
+            Class::Application => "APPLICATION",
+            Class::ContextSpecific => "CONTEXT-SPECIFIC",
+            Class::Private => "PRIVATE",
+        };
+        write!(f, "{}", s)
+    }
+}
+
+impl TryFrom<u8> for Class {
+    type Error = ClassFromIntError;
+
+    #[inline]
+    fn try_from(value: u8) -> Result<Self, Self::Error> {
+        match value {
+            0b00 => Ok(Class::Universal),
+            0b01 => Ok(Class::Application),
+            0b10 => Ok(Class::ContextSpecific),
+            0b11 => Ok(Class::Private),
+            _ => Err(ClassFromIntError(())),
+        }
+    }
 }
